make ChunkAsyncStreamIterator a generic aiohttp helper
