+++ homeassistant/components/cloud/backup.py
@@ -6,9 +6,9 @@ import base64
 from collections.abc import AsyncIterator, Callable, Coroutine, Mapping
 import hashlib
 import logging
-from typing import Any, Self
+from typing import Any
 
-from aiohttp import ClientError, ClientTimeout, StreamReader
+from aiohttp import ClientError, ClientTimeout
 from hass_nabucasa import Cloud, CloudError
 from hass_nabucasa.cloud_api import (
     async_files_delete_file,
@@ -19,6 +19,7 @@ from hass_nabucasa.cloud_api import (
 
 from homeassistant.components.backup import AgentBackup, BackupAgent, BackupAgentError
 from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.aiohttp_client import ChunkAsyncStreamIterator
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
 from .client import CloudClient
@@ -71,31 +72,6 @@ def async_register_backup_agents_listener(
 
     unsub_signal = async_dispatcher_connect(hass, EVENT_CLOUD_EVENT, handle_event)
     return unsub
-
-
-class ChunkAsyncStreamIterator:
-    """Async iterator for chunked streams.
-
-    Based on aiohttp.streams.ChunkTupleAsyncStreamIterator, but yields
-    bytes instead of tuple[bytes, bool].
-    """
-
-    __slots__ = ("_stream",)
-
-    def __init__(self, stream: StreamReader) -> None:
-        """Initialize."""
-        self._stream = stream
-
-    def __aiter__(self) -> Self:
-        """Iterate."""
-        return self
-
-    async def __anext__(self) -> bytes:
-        """Yield next chunk."""
-        rv = await self._stream.readchunk()
-        if rv == (b"", False):
-            raise StopAsyncIteration
-        return rv[0]
 
 
 class CloudBackupAgent(BackupAgent):
